Rename ProjectResponse into ProjectDto
